homeymind-ui/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pyyaml==6.0.1
sse-starlette==1.8.2 